chore: Extract SimpleDERReader to a utility package. Also introduce the XMLSecModule.
